+++ src/libraries/Microsoft.Bcl.Cryptography/src/System/Security/Cryptography/NetStandardShims.cs
@@ -114,6 +114,7 @@ namespace System.Security.Cryptography
         }
     }
 
+#if !NETSTANDARD2_1_OR_GREATER
     internal static class CryptographicOperations
     {
         [MethodImpl(MethodImplOptions.NoInlining | MethodImplOptions.NoOptimization)]
@@ -146,6 +147,7 @@ namespace System.Security.Cryptography
             return accum == 0;
         }
     }
+#endif
 }
 
 namespace System.Runtime.CompilerServices
