--- src/libraries/System.Security.Cryptography/src/System/Security/Cryptography/PbeEncryptionAlgorithm.cs
@@ -1,14 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-namespace System.Security.Cryptography
-{
-    public enum PbeEncryptionAlgorithm
-    {
-        Unknown = 0,
-        Aes128Cbc = 1,
-        Aes192Cbc = 2,
-        Aes256Cbc = 3,
-        TripleDes3KeyPkcs12 = 4,
-    }
-}
