--- src/libraries/System.Security.Cryptography/src/System/Security/Cryptography/PbeParameters.cs
@@ -1,24 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-namespace System.Security.Cryptography
-{
-    public sealed class PbeParameters
-    {
-        public PbeEncryptionAlgorithm EncryptionAlgorithm { get; }
-        public HashAlgorithmName HashAlgorithm { get; }
-        public int IterationCount { get; }
-
-        public PbeParameters(
-            PbeEncryptionAlgorithm encryptionAlgorithm,
-            HashAlgorithmName hashAlgorithm,
-            int iterationCount)
-        {
-            ArgumentOutOfRangeException.ThrowIfNegativeOrZero(iterationCount);
-
-            EncryptionAlgorithm = encryptionAlgorithm;
-            HashAlgorithm = hashAlgorithm;
-            IterationCount = iterationCount;
-        }
-    }
-}
